Youtube API Addition
Added pulling of youtube videos via Ajax requests of YouTube_ID.php, to
get the newest songs youtube video ID.  It then checks if the video
currently playing as ended. If ended, then the AJAX is polled again to
see if a new song is played, which is then loaded into the youtube
player. Thus creating a video radio via twitter feed.