#ThrowBackThursday Fix
Adds a warning notice to users that alert them that the channel will
repeat all Thursday due to the XM Feed being down every Thursday to
accommodate the takeover of their channel by Utopia (90's music). A
future update which will play a mix of past songs on Thursdays will need
to be added to avoid interruption.