Added Song Time Limit
Some song mixes were coming up at 50+ minutes in length, which ruins the
experience of a radio which should always be changing songs.  This Fix
adds a timeout at 6 minutes, to change the song (Radio.php) or seek the
next song (Live.php) to keep things lively! This way if a song is too
long it will move onto the next one seamlessly.